requirements.txt
pyrogram
tgcrypto
mysql-connector-python
requests
html2text